requirements.txt
Django>=5.2,<5.3
wagtail>=7.2
psycopg2-binary>=2.9
Pillow>=10.0
django-extensions>=3.2
python-decouple>=3.8
gunicorn>=21.2
whitenoise>=6.6
requests>=2.31
django-htmx>=1.17
django-ratelimit>=4.1
